requirements_desktop_capture.txt
# Dependências para Desktop Capture (Comet Vision)
# Instalação: pip install -r requirements_desktop_capture.txt

# Captura de tela e manipulação de imagens
Pillow>=10.0.0

# Informações de processos e sistema
psutil>=5.9.0

# Requisições HTTP
requests>=2.31.0

# Agendamento automático
schedule>=1.2.0

# Windows only (detecção de janelas)
# Descomente se estiver no Windows:
# pywin32>=306
